+++ facade/src/main/java/com/timevale/forward/facade/api/query/ProductDemandQueryList.java
@@ -23,7 +23,7 @@ public class ProductDemandQueryList extends QueryBase {
     private String name;
 
     @ApiModelProperty("优先级:0(P0),1(P1),2(P2),3(P3)")
-    private List<Byte> priorities;
+    private List<Integer> priorities;
 
     @ApiModelProperty("业务域")
     private List<Long> bizDomainIds;
@@ -32,7 +32,7 @@ public class ProductDemandQueryList extends QueryBase {
     private List<Long> productLineIds;
 
     @ApiModelProperty("状态:0待排期,10已列入项目,20项目进行中,30已完成上线,40已暂停,50已作废")
-    private Byte status;
+    private Integer status;
 
     @ApiModelProperty("负责人")
     private List<String> owners;
